% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_QC_metrics.R
\name{calculate_QC_metrics}
\alias{calculate_QC_metrics}
\title{Calculate QC metrics}
\usage{
calculate_QC_metrics(data = NULL, group_by = NULL)
}
\arguments{
\item{data}{A tidyseurat object merged with metadata. Must contain columns
"Well_ID", "Row", "Column".}

\item{group_by}{A metadata column name to group data}
}
\value{
a list with a data frame with QC metrics and a box plot showing read counts per condition
}
\description{
To calculate QC metrics such as standard deviation (sd), median absolute deviation (MAD),
interquartile range (IQR), and Z score
for read counts per condition/group of interest
}
