% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_enrichr_pathways.R
\name{compute_enrichr_pathways}
\alias{compute_enrichr_pathways}
\title{Perform enrichR-style analysis on a screen}
\usage{
compute_enrichr_pathways(
  data,
  genesets = enrichr_genesets,
  species = NULL,
  p_value_cutoff = 0.01,
  n_distinct = 5
)
}
\arguments{
\item{data}{A tidyseurat object merged with metadata. Must contain columns
"Well_ID", "Row", "Column".}

\item{genesets}{Named list of genes}

\item{species}{One of "human", "mouse", "fly", "yeast", "worm" or "fish"}

\item{p_value_cutoff}{Cutoff for adjusted p-value (column p_value_adj), default 0.01}

\item{n_distinct}{Minimum number of genes in a geneset, default 5}
}
\value{
A tidyseurat object with pathway_enrichment list in slot tools
}
\description{
Perform enrichR-style analysis on a screen
}
\examples{
file_path <- system.file("extdata", "PMMSq033/PMMSq033.rds", package = "macpie")
mac <- readRDS(file_path)
file_path <- system.file("extdata", "PMMSq033/pathways.Rds", package = "macpie")
genesets <- readRDS(file_path)
multi_enrich_pathways(mac, genesets = genesets)
}
