% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_multi_de.R
\name{compute_multi_de}
\alias{compute_multi_de}
\title{Perform DE of multiple samples in a screen vs control}
\usage{
compute_multi_de(
  data = NULL,
  treatment_samples = NULL,
  control_samples = NULL,
  method = "edgeR",
  num_cores = 2,
  batch = 1,
  k = 2,
  spikes = NULL
)
}
\arguments{
\item{data}{A tidyseurat object merged with metadata. Must contain columns
"Well_ID", "Row", "Column"}

\item{treatment_samples}{Value in the column "combined_id" representing replicates of treatment samples in the data}

\item{control_samples}{Value in the column "combined_id"  representing replicates of control samples in the data}

\item{method}{One of "Seurat_wilcox", "DESeq2", "edgeR", "RUVg", "RUVs", "RUVr", "limma_voom"}

\item{num_cores}{Number of cores}

\item{batch}{Either empty, a single value, or a vector corresponding to the
number of samples}

\item{k}{Parameter k for RUVSeq methods, check RUVSeq tutorial}

\item{spikes}{List of genes to use as spike controls}
}
\value{
List of DE counts vs control
}
\description{
Perform DE of multiple samples in a screen vs control
}
\examples{
data("mini_mac")
control_samples <- "DMSO_0"
mini_mac$combined_id<-make.names(mini_mac$combined_id)
treatment_samples <- c("Staurosporine_10", "Paclitaxel_10", 
"Chlorambucil_10", "Vinblastine_sulfate_10", "Etoposide_10" , 
"Cytarabine_10", "Camptothecin_10", "Anastrozole_10", 
"Sb590885_10", "Fluvastatin_sodium_10")
mini_mac_test<-compute_multi_de(mini_mac, treatment_samples, 
control_samples, num_cores = 1, method = "edgeR")
}
