% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_multi_enrichr.R
\name{compute_multi_enrichr}
\alias{compute_multi_enrichr}
\title{Perform enrichR-style analysis on a screen}
\usage{
compute_multi_enrichr(
  data,
  genesets = NULL,
  species = NULL,
  direction = "both",
  p_value_cutoff = 0.01,
  n_distinct = 10
)
}
\arguments{
\item{data}{A tidyseurat object merged with metadata. Must contain columns
"Well_ID", "Row", "Column".}

\item{genesets}{Named list of genes}

\item{species}{One of "human", "mouse", "fly", "yeast", "worm" or "fish"}

\item{direction}{Direction of the differentially expressed genes,
one of "up", "down", "both" (default).}

\item{p_value_cutoff}{Cutoff for adjusted p-value (column p_value_adj), default 0.01}

\item{n_distinct}{Minimum number of genes in a geneset, default 5}
}
\value{
A tidyseurat object with appended pathway_enrichment dataframe in slot tools
}
\description{
Perform enrichR-style analysis on a screen
}
\examples{
file_path <- system.file("extdata", "PMMSq033/PMMSq033.rds", package = "macpie")
mac <- readRDS(file_path)
file_path <- system.file("extdata", "PMMSq033/pathways.Rds", package = "macpie")
genesets <- readRDS(file_path)
compute_multi_enrichr(mac, genesets = genesets)
}
