% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_multi_screen_profile.R
\name{compute_multi_screen_profile}
\alias{compute_multi_screen_profile}
\title{Find similarities between expression profiles with fgsea}
\usage{
compute_multi_screen_profile(
  data = NULL,
  target = NULL,
  geneset = NULL,
  n_genes_profile = 200,
  direction = "both",
  num_cores = parallel::detectCores() - 1
)
}
\arguments{
\item{data}{A tidyseurat object merged with metadata. Must contain columns
"Well_ID", "Row", "Column".}

\item{target}{Value of a target expression profile in data (target) against
which the others should be compared to}

\item{geneset}{List of genes whose enrichment in expression profiles should
be evaluated}

\item{n_genes_profile}{Number of genes for a target profile, default 200}

\item{direction}{Direction of the expression of DE genes in the target
profile: one of "up", "down" and "both", default both.}

\item{num_cores}{Number of cores}
}
\value{
A tidyseurat object with screen_profile data frame in slot tools
}
\description{
Find similarities between expression profiles with fgsea
}
\examples{
library(macpie)
file_path <- system.file("extdata", "PMMSq033/PMMSq033.rds", package = "macpie")
mac <- readRDS(file_path)
mac <- multi_screen_profile(mac, target = "Staurosporine_10",
n_genes_profile = 100, direction = "up", num_cores = 1)
}
