% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_single_dose_response.R
\name{compute_single_dose_response}
\alias{compute_single_dose_response}
\title{Model Gene Dose-Response Curve Using drc}
\usage{
compute_single_dose_response(
  data,
  gene,
  normalisation = "cpm",
  treatment_value,
  batch = 1
)
}
\arguments{
\item{data}{A Seurat or TidySeurat object containing expression and metadata.}

\item{gene}{A gene name (must match a row name in the object).}

\item{normalisation}{One of "raw", "logNorm", "cpm", "clr", "SCT", "DESeq2",
"edgeR", "RUVg", "RUVs", "RUVr", "limma_voom", "zinb". If empty, defaults to cpm}

\item{treatment_value}{A character string matching one value in metadata column "Treatment_1".}
}
\value{
A list with drc model, predicted values, and ggplot curve
}
\description{
Model Gene Dose-Response Curve Using drc
}
