% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_single_enrichr.R
\name{compute_single_enrichr}
\alias{compute_single_enrichr}
\title{Pathway enrichment analysis}
\usage{
compute_single_enrichr(
  genes = NULL,
  db = NULL,
  genesets = NULL,
  species = NULL
)
}
\arguments{
\item{genes}{Differentially expressed genes for the enrichment analysis}

\item{db}{Valid name of an enrichR database}

\item{genesets}{Named list of genes}

\item{species}{One of "human", "mouse", "fly", "yeast", "worm" or "fish"}
}
\value{
Data frame of pathway-enrichment statistics
}
\description{
pathway_enrichment performs hypergeometric test of enrichment of a set of genes
in user-provided genesets of enricher databases
}
\examples{
file_path <- system.file("extdata", "PMMSq033/PMMSq033.rds", package = "macpie")
mac <- readRDS(file_path)
treatment_samples="Staurosporine_0.1"
control_samples<-"DMSO_0"
top_table <- compute_single_de(mac, treatment_samples, control_samples, method = "limma_voom")
top_genes <- top_table$gene[top_table$p_value_adj<0.01]
result <- compute_single_enrichr(genes = top_genes, db = "MSigDB_Hallmark_2020", species = "human")
head(result)
}
