% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_smiles.R
\name{compute_smiles}
\alias{compute_smiles}
\title{Annotate tidyseurat object with SMILES}
\usage{
compute_smiles(data)
}
\arguments{
\item{data}{A tidyseurat object with a column named \code{Treatment_1}.}
}
\value{
A tidyseurat object with a \code{smiles} column added to the metadata.
}
\description{
This function retrieves isomeric SMILES from PubChem based on compound names
in the \code{Treatment_1} column of a tidyseurat object.
}
\examples{
file_path <- system.file("extdata", "PMMSq033/PMMSq033.rds", package = "macpie")
mac <- readRDS(file_path)
mac <- compute_smiles(mac)
}
