% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_human_to_mouse_ids.R
\name{convert_human_to_mouse}
\alias{convert_human_to_mouse}
\title{For a given gene set of human symbols return mouse symbols}
\usage{
convert_human_to_mouse(gene_list)
}
\arguments{
\item{gene_list}{List of human genes to be converted to mouse IDs}
}
\value{
genes
}
\description{
For a given gene set of human symbols return mouse symbols
}
\examples{
\donttest{
convert_human_to_mouse(c("BRCA1", "TRAF1", "MYBL1"))
}
}
