% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_geneset.R
\name{download_geneset}
\alias{download_geneset}
\title{Title}
\usage{
download_geneset(species = "human", db = "MSigDB_Hallmark_2020")
}
\arguments{
\item{species}{One of "human", "mouse", "fly", "yeast", "worm" or "fish"}

\item{db}{Valid name of an enrichR database}
}
\value{
List of genes per geneset
}
\description{
Title
}
\examples{
genesets <- download_geneset("human", "MSigDB_Hallmark_2020")
head(genesets[["Adipogenesis"]])
}
