% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{enrichr_url}
\alias{enrichr_url}
\title{Connect to enrichR
This function connect to enrichR database
here, we only focus on main enrichR, not fish/fly/yeastenrichR
enrichR contains both human and mouse}
\usage{
enrichr_url(dbs = "enrichr")
}
\arguments{
\item{dbs}{name of database in encirhR such as main enrichR or fish/fly/yeastenrichR}
}
\value{
link for enrichR database
}
\description{
Connect to enrichR
This function connect to enrichR database
here, we only focus on main enrichR, not fish/fly/yeastenrichR
enrichR contains both human and mouse
}
