% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_genes_by_expression.R
\name{filter_genes_by_expression}
\alias{filter_genes_by_expression}
\title{Filter genes by expression and grouping}
\usage{
filter_genes_by_expression(
  data,
  group_by = "combined_id",
  min_counts = 10,
  min_samples = 2
)
}
\arguments{
\item{data}{A tidyseurat object merged with metadata. Must contain columns
"Well_ID", "Row", "Column"}

\item{group_by}{Name of the column that defines groups of replicates}

\item{min_counts}{Minimum number of reads per gene per group}

\item{min_samples}{Minimum number of samples in a group}
}
\value{
tidyseurat object
}
\description{
Filter genes by expression and grouping
}
\examples{
file_path <- system.file("extdata", "PMMSq033/PMMSq033.rds", package = "macpie")
mac <- readRDS(file_path)
mac <- filter_genes_by_expression(mac, group_by = "combined_id", min_counts = 10, min_samples = 2)
}
