% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_clusters_de_umap.R
\name{find_clusters_de_umap}
\alias{find_clusters_de_umap}
\title{Calculate clusters for umap based on DE analysis}
\usage{
find_clusters_de_umap(data = NULL, k = 10)
}
\arguments{
\item{data}{tidyseurat object merged with metadata. Must contain columns
"Well_ID", "Row", "Column".}

\item{k}{Number of nearest neighbors in buildSNNGraph used to construct a
shared nearest-neighbor (SNN) graph}
}
\value{
A tidyseurat object with cluster information in the metadata slot
}
\description{
Calculate clusters for umap based on DE analysis
}
\examples{
data(mini_mac)
mini_mac <- compute_de_umap(mini_mac)
mini_mac <- find_clusters_de_umap(mini_mac)
}
