% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findmetadata.R
\name{findmetadata}
\alias{findmetadata}
\title{User interactive function to find the metadata file}
\usage{
findmetadata(
  predefined_columns = c("Plate_ID", "Well_ID", "Row", "Column", "Species", "Cell_type",
    "Model_type", "Time", "Unit", "Treatment_1", "Concentration_1", "Unit_1",
    "Sample_type", "Barcode")
)
}
\arguments{
\item{predefined_columns}{for findmetadata function, we are string matching the 14 columns from user input metadatafile}
}
\value{
it returns a dataframe of metadafile
}
\description{
This is a rigid function compared to findmetadata_liberal

The function ask user to select a metadatafile

It checks the read permission of the file and its format

Then it will read the user file and check column headers agaisnt the predefined columns

It will also provide a quick summary of matched columns identified

User should ideally have 14 matched columns
}
\examples{
ML is too lazy to write
meta_dt <- findmetadata()
}
\author{
Mark Li
}
