% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findmetadata_liberal.R
\name{findmetadata_liberal}
\alias{findmetadata_liberal}
\title{User interactive function to find the metadata file with freedom
Yet to be decided by the dev team discussion}
\usage{
findmetadata_liberal(
  predefined_columns = c("Plate_ID", "Well_ID", "Row", "Column", "Species", "Cell_type",
    "Model_type", "Time", "Unit", "Treatment_1", "Concentration_1", "Unit_1",
    "Sample_type", "Barcode")
)
}
\arguments{
\item{predefined_columns}{this is the most important and rigid input for the entire package}
}
\value{
it returns a dataframe of metadafile
}
\description{
This is a liberal function compared to findmetadata

The function ask user to select a metadatafile

It checks the read permission of the file and its format

Then it will read the user file and check column headers

Then it will ask user to identify what column of the input file matches the predefined column one by one

ML can also modify the matching process to use a number vector

User should ideally have 14 matched columns
}
\examples{
testmetadata <- findmetadata_liberal()
}
\author{
Mark Li
}
