% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{geneset_download}
\alias{geneset_download}
\title{Download dataset of interest}
\usage{
geneset_download(url, geneset)
}
\arguments{
\item{url}{link of enrichR database}

\item{geneset}{name of data sets from the enrichR database such as KEGG}
}
\value{
list of pathways contains sub-lists of genes that annotated in the corresponding pathways
}
\description{
Download dataset of interest
}
