% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets.R
\name{genesets}
\alias{genesets}
\title{Pathway gene‐sets for PMMSq033}
\format{
A named list of length character vectors. Each element is
a character vector of gene symbols corresponding to one pathway.
}
\source{
Generated from
\code{extdata/PMMSq033/pathways.Rds}
via the \code{data-raw/data.R} preprocessing script.
}
\description{
A named list of gene sets used for hypergeometric enrichment tests
in the PMMSq033 dataset.
}
