% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{GS_summary}
\alias{GS_summary}
\title{2.check enrichr connection and get the stats of datasets from enrichR
This function returns a data frame with all available data sets from enrichR
from enrichR script, the listEnrichrSites function extracts the big table of all available datasets
from "datasetStatistics"}
\usage{
GS_summary(url)
}
\arguments{
\item{url}{link of the enrichR database}
}
\value{
data frame of all available data sets from the enrichR link user provided
}
\description{
2.check enrichr connection and get the stats of datasets from enrichR
This function returns a data frame with all available data sets from enrichR
from enrichR script, the listEnrichrSites function extracts the big table of all available datasets
from "datasetStatistics"
}
