% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyper_enrich_bg.R
\name{hyper_enrich_bg}
\alias{hyper_enrich_bg}
\title{Calculate enrichment of DE genes in a gene set
Internal function}
\usage{
hyper_enrich_bg(deg = NULL, genesets = NULL, background = "human")
}
\arguments{
\item{deg}{vector of differentially expressed genes}

\item{genesets}{list of genes per pathway from enrichr}

\item{background}{defaults to the number of genes in the geneset, otherwise a
number of genes (integer) or species}
}
\value{
enricment stats
}
\description{
Calculate enrichment of DE genes in a gene set
Internal function
}
\examples{
file_path <- system.file("extdata", "PMMSq033/PMMSq033.rds", package = "macpie")
mac <- readRDS(file_path)
treatment_samples="Staurosporine_0.1"
control_samples<-"DMSO_0"
top_table <- differential_expression(mac, treatment_samples, control_samples,
method = "limma_voom")
top_genes <- top_genes<-top_table \%>\%
  filter(.data$p_value_adj<0.01) \%>\%
    select(gene) \%>\%
    pull()
genesets <- geneset_download(url=paste0("https://maayanlab.cloud/", NULL,
"Enrichr/"), "MSigDB_Hallmark_2020")
hyper_enrich_bg(top_genes, genesets) \%>\% head()
}
