% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_clusters_de_umap.R
\name{multi_cluster_umap}
\alias{multi_cluster_umap}
\title{Calculate clusters for umap based on DE analysis}
\usage{
multi_cluster_umap(data = NULL, k = 10)
}
\arguments{
\item{data}{tidyseurat object merged with metadata. Must contain columns
"Well_ID", "Row", "Column".}

\item{k}{Number of nearest neighbors in buildSNNGraph used to construct a
shared nearest-neighbor (SNN) graph}
}
\value{
A tidyseurat object with cluster information in the metadata slot
}
\description{
Calculate clusters for umap based on DE analysis
}
\examples{
file_path <- system.file("extdata", "PMMSq033/PMMSq033.rds", package = "macpie")
mac <- readRDS(file_path)
mac <- multi_prepare_umap(mac)
mac <- multi_cluster_umap(mac)
multi_plot_umap(mac, group_by = "cluster", max_overlaps = 10)
}
