% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_DE.R
\name{multi_DE}
\alias{multi_DE}
\title{Perform DE of multiple samples in a screen vs control}
\usage{
multi_DE(
  data = NULL,
  treatment_samples = NULL,
  control_samples = NULL,
  method = "edgeR",
  num_cores = NULL,
  batch = 1,
  k = 2,
  spikes = NULL
)
}
\arguments{
\item{data}{A tidyseurat object merged with metadata. Must contain columns
"Well_ID", "Row", "Column"}

\item{treatment_samples}{Value in the column "combined_id" representing replicates of treatment samples in the data}

\item{control_samples}{Value in the column "combined_id"  representing replicates of control samples in the data}

\item{method}{One of "Seurat_wilcox", "DESeq2", "edgeR", "RUVg", "RUVs", "RUVr", "limma_voom"}

\item{num_cores}{}

\item{batch}{Either empty, a single value, or a vector corresponding to the
number of samples}

\item{k}{Parameter k for RUVSeq methods, check RUVSeq tutorial}

\item{spikes}{List of genes to use as spike controls}
}
\description{
Perform DE of multiple samples in a screen vs control
}
