% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cpm.R
\name{plot_cpm}
\alias{plot_cpm}
\title{Generate a box plot to show gene expression (CPM)
This is the function to generate a box plot to show CPM levels of DE genes
among selected treatment samples and control samples.}
\usage{
plot_cpm(
  data = NULL,
  genes = NULL,
  group_by = NULL,
  treatment_samples = NULL,
  control_samples = NULL
)
}
\arguments{
\item{data}{A tidyseurat object merged with metadata. Must contain columns
"Well_ID", "Row", "Column".}

\item{genes}{Genes to be plotted}

\item{group_by}{A column that specifies the treatment group in the input data}

\item{treatment_samples}{Value in the column "combined_id" representing replicates of treatment samples in the data}

\item{control_samples}{Value in the column "combined_id"  representing replicates of control samples in the data}
}
\value{
a ggplot2 object
}
\description{
Generate a box plot to show gene expression (CPM)
This is the function to generate a box plot to show CPM levels of DE genes
among selected treatment samples and control samples.
}
\examples{
file_path <- system.file("extdata", "PMMSq033/PMMSq033.rds", package = "macpie")
mac <- readRDS(file_path)
treatment_samples <- "Staurosporine_0.1"
control_samples <- "DMSO_0"
group_by <- "combined_id"
top_table_edgeR <- compute_single_de(mac, treatment_samples, control_samples, method = "edgeR")
genes <- c("TBRG4", "MRPL52", "DCTPP1", "ZFP36L1", "LSM1", "POLR2G")
plot_cpm(mac,genes, group_by, treatment_samples, control_samples)
}
