% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_de_umap.R
\name{plot_de_umap}
\alias{plot_de_umap}
\title{Plot UMAP dimensionality reduction on DE genes}
\usage{
plot_de_umap(data = NULL, color_by = NULL, label = NULL, max_overlaps = NULL)
}
\arguments{
\item{data}{A tidyseurat object merged with metadata. Must contain columns
"Well_ID", "Row", "Column"}

\item{color_by}{A string specifying which column in data will be used to
color the samples.}

\item{label}{A string specifying which column in data will be used to
label a sample.}

\item{max_overlaps}{Maximum number of overlaps for ggrepel}
}
\value{
ggplot object
}
\description{
Plot UMAP dimensionality reduction on DE genes
}
