% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{plot_enrich}
\alias{plot_enrich}
\title{generate a basic bar plot}
\usage{
plot_enrich(df, num_terms = 10)
}
\arguments{
\item{df}{output data frame from enrichment analysis}

\item{num_terms}{number of pathway terms to be plotted}
}
\value{
bar plot showing top 10 enriched pathways
}
\description{
generate a basic bar plot
}
