% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_ranks.R
\name{plot_gene_ranks}
\alias{plot_gene_ranks}
\title{Generate a knee plot
A knee plot to show total number of total read counts for each gene in a given
treatment group}
\usage{
plot_gene_ranks(data = NULL, group_by = NULL, samples = NULL)
}
\arguments{
\item{data}{A tidyseurat object merged with metadata. Must contain columns
"Well_ID", "Row", "Column".}

\item{group_by}{A column that specifies the treatment group in the input data}

\item{samples}{Treatment group}
}
\description{
Generate a knee plot
A knee plot to show total number of total read counts for each gene in a given
treatment group
}
\examples{
file_path <- system.file("extdata", "PMMSq033/PMMSq033.rds", package = "macpie")
mac <- readRDS(file_path)
plot_rank_counts(mac, group_by = "combined_id", samples = "Staurosporine_10")
}
