% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_plate_layout.R
\name{plot_plate_layout}
\alias{plot_plate_layout}
\title{Plot MAC-seq data on a plate layout}
\usage{
plot_plate_layout(
  data = NULL,
  metric = NULL,
  annotation = NULL,
  midpoint = NULL
)
}
\arguments{
\item{data}{A tidyseurat object merged with metadata. Must contain columns
"Well_ID", "Row", "Column"}

\item{metric}{A string specifying which column in data will be used to color
a sample. Defaults to "nCount_RNA".}

\item{annotation}{A string specifying which column in data will be used to
label a sample. Defaults to "Treatment_1".}

\item{midpoint}{A value to be used in heatmap scale bar, it can be either mean value or median value.}
}
\value{
ggplot object
}
\description{
Plot MAC-seq data on a plate layout
}
\examples{

data("mini_mac")
p <- plot_plate_layout(mini_mac, metric = "nCount_RNA",
annotation = "Treatment_1", midpoint = "mean")
}
