% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_QC_metrics.R
\name{plot_qc_metrics}
\alias{plot_qc_metrics}
\title{Create a lollipop chart}
\usage{
plot_qc_metrics(stats_summary, group_by, metrics)
}
\arguments{
\item{stats_summary}{stats_summary a list}

\item{group_by}{A metadata column name to group data}

\item{metrics}{to specify one of the QC metrics from stats_summary calculated from QC_metrics}
}
\description{
Plot QC metrics such as standard deviation (sd), median absolute deviation (MAD),
interquartile range (IQR), or Z score
or read counts per condition/group of interest in a lollipop chart
}
