% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qc_metrics_heatmap.R
\name{plot_qc_metrics_heatmap}
\alias{plot_qc_metrics_heatmap}
\title{Create a heatmap for multiple QC metrics}
\usage{
plot_qc_metrics_heatmap(
  stats_summary,
  group_by = NULL,
  metrics = NULL,
  sort_by = NULL
)
}
\arguments{
\item{stats_summary}{A list containing QC metrics.}

\item{group_by}{A metadata column name to group data.}

\item{metrics}{A vector of QC metrics to visualize.}

\item{sort_by}{The metric to use for sorting group_by values.}
}
\description{
Plots multiple QC metrics from stats_summary as a heatmap, normalizing values between 0 and 1.
}
