% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Volcano plot of differentially expressed genes}
\usage{
plot_volcano(
  top_table,
  x = "log2FC",
  y = "p_value_adj",
  fdr_cutoff = 0.05,
  max.overlaps = 30
)
}
\arguments{
\item{top_table}{Data frame with columns that contain FC and P-values, default to log2FC and p_value_adj}

\item{x}{Name of the column with logFC values}

\item{y}{Name of the column with adjusted p-values}

\item{fdr_cutoff}{Cutoff for labels to be plotted based on their adjusted p-values}

\item{max.overlaps}{Maximum number of overlaps of points for package ggrepel}
}
\value{
A ggplot object
}
\description{
Volcano plot of differentially expressed genes
}
\examples{
data("mini_mac")
top_table <- mini_mac@tools$diff_exprs$Staurosporine_10
plot_volcano(top_table = top_table, max.overlaps = 100)

}
