% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{read_metadata}
\alias{read_metadata}
\title{Read Metadata from a File}
\usage{
read_metadata(
  file_path,
  header = TRUE,
  sep = ",",
  string_as_factors = FALSE,
  predefined_columns = c("Plate_ID", "Well_ID", "Row", "Column", "Species", "Cell_type",
    "Model_type", "Time", "Unit", "Treatment_1", "Concentration_1", "Unit_1",
    "Sample_type", "Barcode", "Project")
)
}
\arguments{
\item{file_path}{A character string specifying the path to the metadata file.
The file must be in CSV or Excel format.}

\item{header}{Logical. If \code{TRUE}, the first row of the file is used as column
names. Defaults to \code{TRUE}.}

\item{sep}{A character string specifying the column separator for CSV files.
Defaults to \code{","}.}

\item{string_as_factors}{Logical. If \code{TRUE}, character columns in the data
frame are converted to factors. Defaults to \code{FALSE}.}

\item{predefined_columns}{A character vector of column names to match in the
file. Defaults to a predefined set of column names:
\code{c("Plate_ID", "Well_ID", "Row", "Column", "Species", "Cell_type", "Model_type",
  "Time", "Unit", "Treatment_1", "Concentration_1", "Unit_1", "Sample_type", "Barcode", "Project")}.}
}
\value{
A data frame containing only the matched columns if at least one
predefined column is found. Returns \code{NULL} and prints an error message if
no predefined columns are found.
}
\description{
This function reads metadata from a specified file, validates the file
format, and extracts columns that match a predefined set of column names.
}
\details{
The function first checks if the file exists and validates its
format (CSV or Excel). It then attempts to read the file and match its
column names to a predefined set of column names. Supported file formats
include:
\itemize{
\item CSV files with \code{.csv} extension.
\item Excel files with \code{.xls} or \code{.xlsx} extensions (requires the \code{readxl} package).
}
}
\examples{
# Example CSV file
file_path <- system.file("/extdata/PMMSq033_metadata.csv", package = "macpie")
result <- read_metadata(file_path)

}
