% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rle_plot.R
\name{rle_plot}
\alias{rle_plot}
\title{Create an RLE Plot}
\usage{
rle_plot(count_matrix, id, feature, logged = FALSE)
}
\arguments{
\item{count_matrix}{A numeric matrix or data frame containing the expression data.
Each column represents a sample, and each row represents a feature.}

\item{id}{A vector of sample identifiers corresponding to the columns of \code{count_matrix}.}

\item{feature}{A vector of feature labels corresponding to the columns of \code{count_matrix}.}

\item{logged}{A logical value indicating whether \code{count_matrix} is already log-transformed.
Defaults to \code{FALSE}.}
}
\value{
A ggplot object representing the RLE plot.
}
\description{
This function generates a Relative Log Expression (RLE) plot for visualizing
the distribution of expression data after normalization or log transformation.
}
\details{
The function performs the following steps:
\itemize{
\item Ensures input data is numeric and matches dimensions.
\item Optionally log-transforms the data if \code{logged = FALSE}.
\item Computes the RLE by subtracting the row medians from each value.
\item Creates a boxplot using ggplot2 to visualize the distribution of RLE values.
}
}
\examples{
# Example Data
rds_file<-system.file("/extdata/PMMSq033/PMMSq033.rds", package = "macpie")
mac<-readRDS(rds_file)
count_matrix<-as.matrix(mac@assays$RNA$counts)
colnames(count_matrix)<-mac$Well_ID
rle_plot(count_matrix = count_matrix, id = mac$Well_ID, feature = mac$Row, logged=FALSE)


}
