% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mofa_macpie.R
\name{run_mofa_macpie}
\alias{run_mofa_macpie}
\title{Run wrapper function for MOFA2}
\usage{
run_mofa_macpie(object, outfile = NULL, save_data = TRUE)
}
\arguments{
\item{object}{an untrained MOFA object}

\item{outfile}{output file for the model (.hdf5 format). If NULL, a temporary file is created.}

\item{save_data}{logical indicating whether to save the training data in the hdf5 file. This is useful for some downstream analysis (mainly functions with the prefix plot_data), but it can take a lot of disk space.}
}
\value{
A trained MOFA2 object
}
\description{
Wrapper function to train an untrained MOFA object from the package MOFA2. To ensure successful build of the conda environment in the docker container, this wrapper sets use_basilisk = TRUE. Please note that not using the basilisk package will make the function fail.
}
\examples{
file <- system.file("extdata", "test_data.RData", package = "MOFA2")
load(file)
MOFAmodel <- create_mofa(dt)
MOFAmodel <- prepare_mofa(MOFAmodel)
run_mofa_macpie(object = MOFAmodel, outfile = NULL, save_data = TRUE)
}
