% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_de.R
\name{summarise_de}
\alias{summarise_de}
\title{Generate a table to summarise gene numbers from a differential
expression test.}
\usage{
summarise_de(
  top_table,
  lfc_threshold = 1,
  padj_threshold = 0.01,
  multi = FALSE,
  group_by = "combined_id"
)
}
\arguments{
\item{top_table}{A data table showing results from compute_single_de}

\item{lfc_threshold}{Threshold of log2FC}

\item{padj_threshold}{Threshold of adjusted p value}

\item{multi}{To indicate to summarise for single de comparison or multi de comparison}

\item{group_by}{Name of the column that defines groups of replicates}
}
\value{
a data table
}
\description{
Generate a table to summarise gene numbers from a differential
expression test.
}
\examples{
data("mini_mac")
top_table <- mini_mac@tools$diff_exprs$Staurosporine_10
summarise_de(top_table, lfc_threshold = 1, padj_threshold = 0.05)
summarise_de(mini_mac, lfc_threshold = 1, 
padj_threshold = 0.01, multi=TRUE)
}
