% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validate.R
\name{validate_macpie_input}
\alias{validate_macpie_input}
\title{Validate macpie input object
Accepts Seurat, tidyseurat, or plain data.frame/tibble (easily extendable).}
\usage{
validate_macpie_input(data, required = NULL)
}
\description{
Validate macpie input object
Accepts Seurat, tidyseurat, or plain data.frame/tibble (easily extendable).
}
\keyword{internal}
