% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmetedata.R
\name{validate_metadata}
\alias{validate_metadata}
\title{Check and clean the metadata file generated from findmetadata function}
\usage{
validate_metadata(metadata)
}
\arguments{
\item{metadata}{this is the output from findmetadata function}
}
\value{
it prints out a summary
}
\description{
As said in title

It checks the column completeness and removes special characters such as spaces and comma except from Treatment_1

It replaces special characters such as spaces and comma with underscore

It generates a quick summary of the different treatment and sample groups in the metadata
}
\examples{
ML is too lazy to write
metadt_qc <- validate_metadata(meta_dt)

}
\author{
Mark Li
}
