% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_by_de.R
\name{aggregate_by_de}
\alias{aggregate_by_de}
\title{Collapse Replicates by Differential Expression}
\usage{
aggregate_by_de(data, metric_col = "metric")
}
\arguments{
\item{data}{A tidyseurat object with DE results}

\item{metric_col}{A character string indicating the column name in each
DE dataframe to use for aggregation. Defaults to "metric".}
}
\value{
A new tidyseurat object with collapsed metadata
}
\description{
This function collapses replicate samples in a tidyseurat object based on
differential expression (DE) results. It returns a new tidyseurat object
with a 'DE' assay containing the aggregated values.
}
\details{
This function performs the following steps:
\itemize{
\item Converts the list of DE comparisons into a gene-by-condition matrix.
\item Aggregates metadata across replicates by \code{combined_id}.
\item Creates a new Seurat object with this matrix as a "DE" assay
}
}
\examples{
file_path <- system.file("extdata", "PMMSq033/PMMSq033.rds", package = "macpie")
mac <- readRDS(file_path)
mac_collapsed <- aggregate_by_de(mac)

}
