% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_chem_descriptors.R
\name{compute_chem_descriptors}
\alias{compute_chem_descriptors}
\title{Compute chemical descriptors from SMILES}
\usage{
compute_chem_descriptors(data)
}
\arguments{
\item{data}{A tidyseurat object with a \code{smiles} column and \code{Treatment_1} column.}
}
\value{
The same tidyseurat object with a new entry in \code{tools$chem_descriptors}.
}
\description{
This function parses SMILES strings and computes chemical descriptors using rcdk.
It stores cleaned, non-redundant descriptors in \code{tools$chem_descriptors}.
}
\examples{
file_path <- system.file("extdata", "PMMSq033/PMMSq033.rds", package = "macpie")
mac <- readRDS(file_path)
mac <- compute_smiles(mac)
mac <- compute_chem_descriptors(mac)
}
