% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_chem_descriptors.R
\name{compute_chem_descriptors}
\alias{compute_chem_descriptors}
\title{Compute chemical descriptors from SMILES}
\usage{
compute_chem_descriptors(
  data,
  treatment_ids = NULL,
  r_squared = 0.6,
  descriptors = NULL
)
}
\arguments{
\item{data}{A tidyseurat object with a \code{smiles} column and \code{Treatment_1} column.}

\item{treatment_ids}{A list of unique sample identifiers, commonly combined_ids}

\item{r_squared}{R squared value, default of 0.6}

\item{descriptors}{Specify a subset of descriptors of interest from rcdk}
}
\value{
The same tidyseurat object with a new entry in \code{tools$chem_descriptors}.
}
\description{
This function parses SMILES strings and computes chemical descriptors using rcdk.
It stores cleaned, non-redundant descriptors in \code{tools$chem_descriptors}.
}
\examples{
\dontrun{
mock_data <- tibble::tibble(
  Treatment_1 = c("Aspirin", "Caffeine", "NonExistentCompound_123")
)
result <- compute_smiles(mock_data,compound_column = "Treatment_1" )
data <- compute_chem_descriptors(result, treatment_ids = mock_data$Treatment_1, descriptors =
"org.openscience.cdk.qsar.descriptors.molecular.FractionalCSP3Descriptor")
}
}
