% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_de_umap.R
\name{compute_de_umap}
\alias{compute_de_umap}
\title{Prepare DE-based umap}
\usage{
compute_de_umap(data = NULL)
}
\arguments{
\item{data}{A tidyseurat object merged with metadata. Must contain columns
"Well_ID", "Row", "Column".}
}
\value{
A tidyseurat object with umap_de data frame in slot tools
}
\description{
Prepare DE-based umap
}
\examples{
file_path <- system.file("extdata", "PMMSq033/PMMSq033.rds", package = "macpie")
mac <- readRDS(file_path)
mac <- compute_de_umap(mac)

}
