% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_hyper_enrich_bg.R
\name{compute_hyper_enrich_bg}
\alias{compute_hyper_enrich_bg}
\title{Calculate enrichment of DE genes in a gene set
Compute one‐sided hypergeometric p-values, z-scores, and combined scores
for a list of differentially expressed genes (\code{deg}) against each entry in
a list of gene sets (\code{genesets}).
Internal function}
\usage{
compute_hyper_enrich_bg(deg = NULL, genesets = NULL, background = "human")
}
\arguments{
\item{deg}{vector of differentially expressed genes}

\item{genesets}{list of genes per pathway from enrichr}

\item{background}{integer or \code{character} (\code{"human"}, \code{"mouse"}, or
\code{"geneset"}).  If \code{"human"}, uses an approximate universe of
\strong{20 000} genes; if \code{"mouse"}, \strong{22 000}; if \code{"geneset"}, the
total unique genes in \code{genesets}.  You may also supply a custom
integer (\code{n}) equal to the exact size of your tested universe.}
}
\value{
enrichment stats
}
\description{
Calculate enrichment of DE genes in a gene set
Compute one‐sided hypergeometric p-values, z-scores, and combined scores
for a list of differentially expressed genes (\code{deg}) against each entry in
a list of gene sets (\code{genesets}).
Internal function
}
\examples{
data(mini_mac)
treatment_samples="Staurosporine_0.1"
control_samples<-"DMSO_0"
top_table <- compute_single_de(mini_mac, treatment_samples, control_samples,
method = "limma_voom")
top_genes <- top_table$gene[top_table$p_value_adj<0.01]
data(genesets)
results <- compute_hyper_enrich_bg(top_genes, genesets)
head(results)
}
