% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_single_dose_response.R
\name{compute_single_dose_response}
\alias{compute_single_dose_response}
\title{Model Gene Dose-Response Curve Using drc}
\usage{
compute_single_dose_response(
  data,
  gene = NULL,
  pathway = NULL,
  normalisation = "limma_voom",
  treatment_value,
  control_value = "DMSO",
  batch = 1,
  k = 2
)
}
\arguments{
\item{data}{A Seurat or TidySeurat object containing expression and metadata.}

\item{gene}{A gene name (must match a row name in the object).}

\item{pathway}{A character string present in the list of enriched pathways.}

\item{normalisation}{One of "raw", "logNorm", "cpm", "clr", "SCT", "DESeq2",
"edgeR", "RUVg", "RUVs", "RUVr", "limma_voom", "zinb". If empty, defaults to cpm}

\item{treatment_value}{A character string matching one value in metadata column "Treatment_1".}

\item{control_value}{A character string matching one value in metadata column "Treatment_1".}

\item{batch}{Either empty, a single value, or a vector corresponding to the
number of samples}

\item{k}{Parameter k for RUVSeq methods, check RUVSeq tutorial}
}
\value{
A list with drc model, predicted values, and ggplot curve
}
\description{
Model Gene Dose-Response Curve Using drc
}
\examples{
\donttest{
data(mini_mac)
res <- compute_single_dose_response(data = mini_mac,
gene = "FSHR",
normalisation = "limma_voom",
treatment_value = "Camptothecin")
res$plot
res <- compute_single_dose_response(data = mini_mac,
pathway = "p53 Pathway",
treatment_value = "Camptothecin")
res$plot
}
}
