% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_single_enrichr.R
\name{compute_single_enrichr}
\alias{compute_single_enrichr}
\title{Pathway enrichment analysis}
\usage{
compute_single_enrichr(
  genes = NULL,
  db = NULL,
  genesets = NULL,
  species = NULL
)
}
\arguments{
\item{genes}{Differentially expressed genes for the enrichment analysis}

\item{db}{Valid name of an enrichR database}

\item{genesets}{Named list of genes}

\item{species}{One of "human", "mouse", "fly", "yeast", "worm" or "fish"}
}
\value{
Data frame of pathway-enrichment statistics
}
\description{
pathway_enrichment performs hypergeometric test of enrichment of a set of genes
in user-provided genesets of enricher databases
}
\examples{
data(mini_mac)
top_genes <- mini_mac@tools$diff_exprs[["Staurosporine_10"]]$gene
result <- compute_single_enrichr(genes = top_genes, db = "MSigDB_Hallmark_2020", species = "human")
head(result)
}
