% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_smiles.R
\name{compute_smiles}
\alias{compute_smiles}
\title{Annotate tidyseurat object with SMILES}
\usage{
compute_smiles(data, compound_column)
}
\arguments{
\item{data}{A tidyseurat object}

\item{compound_column}{Column with the generic name of a compound}
}
\value{
A tidyseurat object with a \code{smiles} column added to the metadata.
}
\description{
This function retrieves isomeric SMILES from PubChem based on compound names
in the \code{Treatment_1} column of a tidyseurat object.
}
\examples{
mock_data <- tibble::tibble(
Treatment_1 = c("Aspirin", "Caffeine", "NonExistentCompound_123")
)
result <- compute_smiles(mock_data, compound_column = "Treatment_1")
}
