% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{hyper_enrich_bg}
\alias{hyper_enrich_bg}
\title{Perform gene enrichment analysis
by doing a hyper-geometric test}
\usage{
hyper_enrich_bg(deg, genesets, background = "human")
}
\arguments{
\item{deg}{list of DEGs}

\item{genesets}{selected pathway datasets like KEGG from enrichR}

\item{background}{number of genes that considered as background, options as "human": number of human genes
or "geneset" genes in the previously selected pathway dataset}
}
\value{
a data frame of enrichment analysis results with all the statistics
}
\description{
Perform gene enrichment analysis
by doing a hyper-geometric test
}
