% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyper_enrich_bg.R
\name{hyper_enrich_bg}
\alias{hyper_enrich_bg}
\title{Calculate enrichment of DE genes in a gene set
Internal function}
\usage{
hyper_enrich_bg(deg = NULL, genesets = NULL, background = "human")
}
\arguments{
\item{background}{}
}
\description{
Calculate enrichment of DE genes in a gene set
Internal function
}
