% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_heatmap.R
\name{metadata_heatmap}
\alias{metadata_heatmap}
\title{Generate Heatmaps of Metadata Function}
\usage{
metadata_heatmap(metadata_file = NULL, legend = "show", output_file = NULL)
}
\arguments{
\item{metadata_file}{Either the existing 'cleaned_metadata' file or a metadata CSV file path.}

\item{legend}{A character value ("show" or "none") to control whether to display legends.}

\item{output_file}{A file path and name to save the heatmaps as a graph (png, pdf, or jpg).}
}
\value{
Displays the plot as a ggplot object or saves it as a file (jpg, png, or pdf).
}
\description{
This function generates heatmaps from metadata, either from an existing
\code{cleaned_metadata} object in the global environment or from a CSV file.
}
\examples{
temp_file <- tempfile(fileext = ".csv")
write.csv(
  data.frame(Well_ID = c("A01", "A02", "B01", "B02"),
  Plate_ID = c("P1", "P1","P1", "P1"),
  Compound_ID = c("C1", "C2", "C3", "C4")),
  temp_file, row.names = FALSE
)
metadata_heatmap(temp_file)
}
