% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Volcano plot of differentially expressed genes}
\usage{
plot_volcano(
  top_table,
  x = "log2FC",
  y = "p_value_adj",
  FDR_cutoff = 0.05,
  max.overlaps = 30
)
}
\arguments{
\item{top_table}{Data frame with columns that contain FC and P-values, default to log2FC and p_value_adj}

\item{x}{Name of the column with logFC values}

\item{y}{Name of the column with adjusted p-values}

\item{FDR_cutoff}{Cutoff for labels to be plotted based on their adjusted p-values}

\item{max.overlaps}{Maximum number of overlaps of points for package ggrepel}
}
\value{
ggpplot plot
}
\description{
Volcano plot of differentially expressed genes
}
\examples{
file_path <- system.file("extdata", "PMMSq033/PMMSq033.rds", package = "macpie")
mac <- readRDS(file_path)
treatment_samples="Staurosporine_0.1"
control_samples<-"DMSO_0"
top_table <- differential_expression(mac, treatment_samples, control_samples, method = "limma_voom")
plot_volcano(top_table)
}
