% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_de.R
\name{summarise_de}
\alias{summarise_de}
\title{Generate a table to summarise gene numbers from a differential
expression test.}
\usage{
summarise_de(
  top_table,
  lfc_threshold = 1,
  padj_threshold = 0.01,
  multi = FALSE,
  group_by = "combined_id"
)
}
\arguments{
\item{top_table}{A data table showing results from compute_single_de}

\item{lfc_threshold}{Threshold of log2FC}

\item{padj_threshold}{Threshold of adjusted p value}

\item{multi}{To indicate to summarise for single de comparison or multi de comparison}

\item{group_by}{Name of the column that defines groups of replicates}
}
\value{
a data table
}
\description{
Generate a table to summarise gene numbers from a differential
expression test.
}
\examples{
file_path <- system.file("extdata", "PMMSq033/PMMSq033.rds", package = "macpie")
mac <- readRDS(file_path)
treatment_samples <- "Staurosporine_0.1"
control_samples <- "DMSO_0"
group_by <- "combined_id"
top_table_edgeR <- compute_single_de(mac, treatment_samples, control_samples, method = "edgeR")
summarise_de(top_table_edgeR, lfc_threshold = 1, padj_threshold = 0.05)
summarise_de(mac, lfc_threshold = 1, padj_threshold = 0.01, multi=TRUE)
}
