% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_metadata.R
\name{validate_metadata}
\alias{validate_metadata}
\title{Check and clean the metadata file generated from findmetadata function}
\usage{
validate_metadata(metadata)
}
\arguments{
\item{metadata}{this is the output from findmetadata function}
}
\value{
A list containing: \item{cleaned_metadata}{The cleaned metadata data
frame after validation and modifications.}
\item{summary_table}{A summary table grouped by \code{Plate_ID}, showing the count of unique values
in selected columns.}
}
\description{
It checks the column completeness and removes special characters
such as spaces and comma except from Treatment_1

It replaces special characters such as spaces and comma with
underscore

It generates a quick summary of the different treatment and
sample groups in the metadata
}
\details{
This function validates a metadata dataset, cleans specific fields, checks for
issues, and generates a summary table grouped by \code{Plate_ID}.
}
\examples{
\donttest{
# Example CSV file
file_path <- system.file("/extdata/PMMSq033_metadata.csv", package = "macpie")
metadata <- read_metadata(file_path)
metadt_qc <- validate_metadata(metadata)
}

}
\author{
Mark Li
}
